# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gantt App API",
    "description": "API REST para gestión de proyectos y tareas con diagramas Gantt",
    "contact": {
      "name": "Gantt App Team"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base URL"
    }
  ],
  "paths": {
...
    "/projects/{project}/tasks": {
      "get": {
        "tags": ["Tareas"],
        "summary": "Listar tareas",
        "description": "Lista paginada de tareas de un proyecto",
        "operationId": "b8c7fa0a8f2164c2591a29a17495861c",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start_date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tareas"
          },
          "404": {
            "description": "Proyecto no encontrado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "post": {
        "tags": ["Tareas"],
        "summary": "Crear tarea",
        "description": "Crea una tarea dentro de un proyecto",
        "operationId": "a9d185726323607499e67377ad9ceef9",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["type", "title"],
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "parent_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "task_status_id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "progress": {
                    "type": "integer"
                  },
                  "dependency_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "dependency_type": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada"
          },
          "422": {
            "description": "Error de validación o ciclo detectado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/tasks/{task}": {
      "get": {
        "tags": ["Tareas"],
        "summary": "Detalle de tarea",
        "description": "Devuelve el detalle de una tarea con relaciones",
        "operationId": "2507ad5dcbf3084e44634aa11b15da4e",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la tarea"
          },
          "404": {
            "description": "No encontrada"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "delete": {
        "tags": ["Tareas"],
        "summary": "Eliminar tarea",
        "description": "Eliminación lógica con cascada a descendientes",
        "operationId": "70a8dde6171ddd9371617d15b1b1b750",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea eliminada"
          },
          "422": {
            "description": "Estado no permite eliminación"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "patch": {
        "tags": ["Tareas"],
        "summary": "Actualizar tarea",
        "description": "Actualización parcial de una tarea",
        "operationId": "39485a75925f3070dd18a69d0f480611",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "task_status_id": {
                    "type": "integer"
                  },
                  "progress": {
                    "type": "integer"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "parent_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "dependency_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada"
          },
          "422": {
            "description": "Transición de estado inválida"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/tasks/{task}/restore": {
      "post": {
        "tags": ["Tareas"],
        "summary": "Restaurar tarea",
        "description": "Restaura una tarea eliminada con cascada",
        "operationId": "07565c37b3ebd9a5ed6a01338551a795",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea restaurada"
          },
          "422": {
            "description": "La tarea no está eliminada"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/tasks/bulk-update": {
      "patch": {
        "tags": ["Tareas"],
        "summary": "Actualización masiva",
        "description": "Actualiza múltiples tareas con los mismos datos",
        "operationId": "bde426156395deb0934a2906e6d07d13",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["task_ids", "data"],
                "properties": {
                  "task_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "data": {
                    "properties": {
                      "task_status_id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "progress": {
                        "type": "integer"
                      },
                      "start_date": {
                        "type": "string",
                        "format": "date"
                      },
                      "end_date": {
                        "type": "string",
                        "format": "date"
                      }
                    },
                    "type": "object"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tareas actualizadas"
          },
          "422": {
            "description": "Error de validación"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "description": "Token de autenticación Laravel Sanctum",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}